# docs/swagger.yaml
basePath: /
definitions:
  controller.LoginInput:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  controller.RegisterInput:
    properties:
      name:
        type: string
      password:
        type: string
      phone_number:
        type: string
      secret_code:
        description: 'Opsional: Untuk jadi Admin'
        type: string
    type: object
...
host: kawai-be.vercel.app
info:
  contact:
    name: Dwi Puspa Firdaus
  description: Dokumentasi API Backend untuk Kawai Assistant (Bot WhatsApp & Dashboard).
  title: Kawai Assistant API
  version: 1.0.0
paths:
...
  /api/admin/ban:
    post:
      consumes:
      - application/json
      description: Memblokir atau membuka blokir user
      parameters:
      - description: Target Phone & Action (ban/unban)
        in: body
        name: input
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: '[ADMIN] Ban User'
      tags:
      - Admin
  /api/admin/login:
    post:
      consumes:
      - application/json
      description: Login untuk admin menggunakan nomor HP dan password
      parameters:
      - description: Kredensial Login
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/controller.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: Token JWT dan Role
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid Body
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Password Salah / Bukan Admin
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login Admin
      tags:
      - Auth
  /api/admin/stats:
    get:
      description: Mengambil statistik sederhana sistem
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: '[ADMIN] Lihat Statistik'
      tags:
      - Admin
  /api/admin/users:
    get:
      description: Mengambil semua data user yang terdaftar
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: '[ADMIN] Lihat Users'
      tags:
      - Admin
  /api/auth/google/callback:
    get:
      description: Menerima kode otorisasi Google dan menukar dengan Token
      parameters:
      - description: Authorization Code
        in: query
        name: code
        required: true
        type: string
      - description: User Phone Number
        in: query
        name: state
        required: true
        type: string
      responses:
        "200":
          description: Halaman Sukses HTML
          schema:
            type: string
      summary: Callback Google Auth
      tags:
      - Auth
  /api/auth/google/login:
    get:
      description: Redirect ke halaman Login Google (OAuth2)
      parameters:
      - description: Nomor HP User (State)
        in: query
        name: phone
        required: true
        type: string
      responses:
        "307":
          description: Redirect ke Google
      summary: Login Google
      tags:
      - Auth
...
  /api/dashboard/notes:
    get:
      description: Mendapatkan daftar catatan milik pengguna yang sedang login.
      parameters:
      - description: Keyword Pencarian
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Ambil Catatan User
      tags:
      - Dashboard
  /api/dashboard/notes/{id}/detail:
    get:
      description: Melihat detail catatan beserta hashtagnya
      parameters:
      - description: Note ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Detail Catatan
      tags:
      - Dashboard
  /api/dashboard/reminders:
    get:
      description: Mengambil daftar pengingat yang akan datang.
      parameters:
      - description: Keyword Pencarian
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Ambil Agenda Pengingat
      tags:
      - Dashboard
...
  /api/register:
    post:
      consumes:
      - application/json
      description: Pendaftaran pengguna baru dengan nomor HP, nama, dan password
      parameters:
      - description: Data Registrasi
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/controller.RegisterInput'
      produces:
      - application/json
      responses:
        "200":
          description: Status Registrasi
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Data tidak lengkap / User sudah ada
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Registrasi User Baru
      tags:
      - Auth
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"